changed u32 to u64

+++ async-openai/src/types/chat.rs
@@ -574,7 +574,7 @@ pub struct CreateChatCompletionResponse {
     /// A list of chat completion choices. Can be more than one if `n` is greater than 1.
     pub choices: Vec<ChatChoice>,
     /// The Unix timestamp (in seconds) of when the chat completion was created.
-    pub created: u32,
+    pub created: u64,
     /// The model used for the chat completion.
     pub model: String,
     /// This fingerprint represents the backend configuration that the model runs with.
@@ -645,7 +645,7 @@ pub struct CreateChatCompletionStreamResponse {
     pub choices: Vec<ChatChoiceStream>,
 
     /// The Unix timestamp (in seconds) of when the chat completion was created. Each chunk has the same timestamp.
-    pub created: u32,
+    pub created: u64,
     /// The model to generate the completion.
     pub model: String,
     /// This fingerprint represents the backend configuration that the model runs with.
